--- src/ignitus-HomePage/ignitus-Contributors/Containers/contributorContainer.ts
@@ -1,1 +1,0 @@
-export {};
