--- src/ignitus-HomePage/ignitus-Contributors/Containers/index.ts
@@ -1,1 +1,0 @@
-export {};
