--- src/ignitus-HomePage/ignitus-Contributors/actionTypes.ts
@@ -1,1 +1,0 @@
-export {};
