--- src/ignitus-HomePage/ignitus-Contributors/actions.ts
@@ -1,1 +1,0 @@
-export {};
