--- src/ignitus-HomePage/ignitus-Contributors/index.ts
@@ -1,6 +1,0 @@
-
-import * as components from './Components';
-
-export default {
-  components,
-};
