--- src/ignitus-HomePage/ignitus-Contributors/reducers.ts
@@ -1,1 +1,0 @@
-export {};
