--- src/ignitus-HomePage/ignitus-Contributors/sagas.ts
@@ -1,2 +1,0 @@
-export default function* sagas():IterableIterator<number> {};
-
