--- src/ignitus-HomePage/ignitus-Contributors/selectors.ts
@@ -1,1 +1,0 @@
-export {};
